refactor(fits): align feature layout with conventions
  * move the flat actions.ts into an actions/ folder so every
    feature exposes server actions the same way (matches auth/)
  * relocate CreateFit into ui/ with the other fit components and
    export it from the barrel, keeping components out of the root
