+++ src/features/fits/ui/index.ts
@@ -1,3 +1,4 @@
+export * from './CreateFit';
 export * from './Fits';
 export * from './Fit';
 export * from './FitsItem';
